Split out validity unit tests into specific condition class tests. Add tests for more invalid cases across condition types. Improve tests to ensure that expected message is raised in exceptions.
